Add files via upload
(Julia) Implements the multicut L-shaped algorithm to solve the following scenario: we have three production facilities, shipping to five demand locations, and we need to decide how to much to ship from each facility to each location, with the objective of maximizing expected profits. All costs and selling prices are known at the outset, but demand is only discovered once the product has shipped.